+++ header.php
@@ -1,14 +1,16 @@
 <!DOCTYPE html>
 <html <?php language_attributes(); ?>>
-<head>
-    <meta charset="<?php bloginfo( 'charset' ); ?>">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <?php wp_head(); ?>
-</head>
-<body <?php body_class(); ?>>
-<?php wp_body_open(); ?>
-<div id="page" class="site flex flex-col min-h-screen">
-    
-    <?php get_template_part( 'template-parts/global/site-header' ); ?>
 
-    <div id="content" class="site-content flex-grow">
+<head>
+  <meta charset="<?php bloginfo('charset'); ?>">
+  <meta name="viewport" content="width=device-width, initial-scale=1">
+  <?php wp_head(); ?>
+</head>
+
+<body <?php body_class(); ?>>
+  <?php wp_body_open(); ?>
+  <div id="page" class="site flex flex-col min-h-screen">
+
+    <?php get_template_part('template-parts/global/site-header'); ?>
+
+    <div id="content" class="site-content grow">