+++ inc/setup.php
@@ -51,3 +51,9 @@ if ( ! function_exists( 'civ_setup' ) ) :
   }
 endif;
 add_action( 'after_setup_theme', 'civ_setup' );
+
+/**
+ * Disable Gutenberg Block Editor.
+ */
+add_filter( 'use_block_editor_for_post', '__return_false', 10 );
+add_filter( 'use_block_editor_for_post_type', '__return_false', 10 );