+++ page-templates/homepage.php
@@ -1,0 +1,13 @@
+<?php
+
+/**
+ * Template Name: Homepage
+ *
+ * @package CIV Shows
+ */
+
+get_header();
+
+get_template_part('template-parts/components/home-hero-slider');
+
+get_footer();